feat(restructure-controller): Renamed authService
